+++ src/routes/user/index.ts
@@ -4,16 +4,16 @@ import loginRouter from './login';
 import roleRouter from './role';
 import notificationRouter from './notification';
 import categoryRouter from './category';
-import { saveUser } from '../../services/User';
+import controller from './user.controller';
 import helloWorld from '../hello-world';
 
 const router = Router();
-router.post('/', saveUser);
+router.post('/', controller.saveUser);
 router.patch('/', helloWorld);
 router.delete('/', helloWorld);
 router.use('/:id/role', roleRouter);
-router.use('/:id/notification', notificationRouter);
-router.use('/:id/category', categoryRouter);
+router.use('/notification', notificationRouter);
+router.use('/category', categoryRouter);
 
 router.use('/profile', profileRouter);
 router.use('/login', loginRouter);
